Renaming file using psr-4 rules
