+++ src/invoker/MerchantInvokeWithBody.php
@@ -4,22 +4,23 @@ namespace Blibli\SellerApi\invoker;
 
 use Blibli\SellerApi\client\BlibliMerchantClient;
 use Blibli\SellerApi\request\ApiConfig;
-use Blibli\SellerApi\request\SignatureGenerator;
+use Blibli\SellerApi\request\GenerateSignature;
 use Exception;
 
-class MerchantInvokeWithoutBody
+class MerchantInvokeWithBody
 {
 	/**
-	 * Function for invoke without body
+	 * Function for POST/PUT/PATCH/DELETE
 	 *
 	 * @param int $port
 	 * @param string $http_method
 	 * @param string $url
 	 * @param array $params
+	 * @param array $body
 	 * @param ApiConfig $request
 	 * @return void
 	 */
-	public static function generateGetMethod($port, $http_method, $url, $params, ApiConfig $request)
+	public static function generateMethod($port, $http_method, $url, $params, $body, ApiConfig $request)
 	{
 		if($request->getToken() == '') throw new Exception("Input of [API Token] is empty!");
 		if($request->getSignatureKey() == '') throw new Exception("Input of [API Secret Key] is empty!");
@@ -27,18 +28,19 @@ class MerchantInvokeWithoutBody
 		if($request->getBusinessPartnerCode() == '') throw new Exception("Input of [Business Partner Code] is empty!");
 		if($request->getPlatformName() == '') throw new Exception("Input of [Platform Name] is empty!");
 		if($request->getTimeoutSecond() == '') $request->setTimeoutSecond(15);
-		$signature = new SignatureGenerator();
+		
+		$signature = new GenerateSignature();
 
 		$milliseconds = round(microtime(true) * 1000);
 		$uuid = BlibliMerchantClient::gen_uuid();
 		$urlMeta = explode("/proxy", $url);
 		$urlRaw = $urlMeta[1];
-		
+
 		if (strpos($urlRaw, "/mta") !== FALSE) {
 			$urlRaw = str_replace("/mta", "/mtaapi", $urlRaw);
 		}
 
-		$signature = $signature->generate($milliseconds, $request->getSignatureKey(), "GET", "", "", $urlRaw);
+		$signature = $signature->generate($milliseconds, $request->getSignatureKey(), "POST", json_encode($body), "application/json", $urlRaw);
 
 		$header = array(
 			"Authorization: bearer " . $request->getToken(),
@@ -72,6 +74,7 @@ class MerchantInvokeWithoutBody
 		CURLOPT_RETURNTRANSFER => true,
 		CURLOPT_ENCODING => "",
 		CURLOPT_MAXREDIRS => 10,
+		CURLOPT_POSTFIELDS => json_encode($body), 
 		CURLOPT_TIMEOUT => $request->getTimeoutSecond(),
 		CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
 		CURLOPT_CUSTOMREQUEST => $http_method,
