+++ src/invoker/basic_auth/SellerInvokeWithBody.php
@@ -4,7 +4,7 @@ namespace Blibli\SellerApi\invoker\basic_auth;
 
 use Blibli\SellerApi\client\BlibliSellerBasicAuthClient;
 use Blibli\SellerApi\request\ApiConfig;
-use Blibli\SellerApi\request\SignatureGenerator;
+use Blibli\SellerApi\request\GenerateSignature;
 use Exception;
 
 class SellerInvokeWithBody
@@ -39,7 +39,7 @@ class SellerInvokeWithBody
 		);
 		
 		if ($request->getSignatureKey()) {
-			$signature = new SignatureGenerator();
+			$signature = new GenerateSignature();
 		
 			$milliseconds = round(microtime(true) * 1000);
 			$urlMeta = explode("/proxy", $url);
